Migrated errors to better location + added general error catch
